add final battle tracking & use real game time
- Added the following tracking messages for the final battle:
"Defeated Tier 1"
"Defeated Tier 2"
"Defeated Tier 3"
"Defeated Kefka - gg!"
These run off two new fields in GetSaveDataResponse: response.onTier = {0, 1, 2, 3, 4} and response.gg = {0, 1}.
These fields update based on monster formation and the "gg bit" $009A, which (in battle mode) enables the Kefka dissolving animation.  Because of this it is actually tracked about 4 seconds after the first 'crackow', which is when we stop the timer in races.  It might be worth timing this out and adjusting it in the code.  Also since both the monster formations and the gg-bit are used for other purposes outside of battle, there is a small chance that this gets tracked under other conditions.  If that's a problem we can further constrain it.

- Switch to the true game time, which is stored at 0x021b (in the menu ram).