Correct Crackow delay
Added a -5 second correction to the "Defeated Kefka - gg!" tracking item.  This tracks off of when the dissolve begins which is about 150 frames (5 seconds) after the actual "Crackow".  This correction makes the tracker timing match racing conventions.  (Note that nothing in the tracker is more accurate than the update rate, i.e. +/- 0.5 seconds.)